Improve Board class
Add functions out_of_range and draw;
Stop importing colors and get them on startup.
